Normalize line endings
The git recommendation is that line endings in repositories be LF, which
work natively on Mac and Unix. On Windows, users can configure git to
convert line endings to CRLF on checkout and back to LF on commit.

I've normalized this repo to LF following the recommendations on this
page:
https://help.github.com/articles/dealing-with-line-endings#platform-mac

If you're on Windows, set git to automatically handle line ending
conversion for you by running:
git config --global core.autocrlf true

Git will then convert to and from CRLF behind the scenes and you
shouldn't need to worry about it ever again.